Effective dart.

+++ lib/type/play_histroy.dart
@@ -3,13 +3,25 @@ import 'episodebrief.dart';
 
 class PlayHistory {
   DBHelper dbHelper = DBHelper();
+
+  /// Episdoe title.
   String title;
+
+  /// Episode url
   String url;
-  double seconds;
+
+  /// Play record seconds.
+  int seconds;
+
+  /// Play record count,
   double seekValue;
+
+  /// Listened date.
   DateTime playdate;
+
   PlayHistory(this.title, this.url, this.seconds, this.seekValue,
       {this.playdate});
+
   EpisodeBrief _episode;
   EpisodeBrief get episode => _episode;
